Fast cut-enumeration algorithm for logic networks with <= 64 nodes (using uint64_t to represent cuts) (#161)
* uint64_t-based cut enumeration algorithm

Computation of cut domination is incorrect though and
needs some work.

* Fix dominating cut condition computation

* Use mockturtle formatting conventions

* Allow user to specify cut size

* Print cuts for visual inspection

* Add test cases for fast cut enumeration algorithm

* Extra docs

* Fix clang bug (c++ standard bug actually)

This technically should have compiled, but there's a bug in the standard here.

The standard says that lambdas can only capture variables. And it says that a non-tuple-like structured binding declaration doesn't introduce variables. It introduces names, but those names aren't names of variables.

A tuple-like structured binding declaration, on the other hand, does introduce variables. a and b in auto [a, b] = std::make_tuple(1, 2); are actual reference-typed variables. So they can be captured by a lambda.

Obviously this is not a sane state of affairs, and the committee knows this, so a fix should be forthcoming (though there appears be some disagreement over exactly how capturing a structured binding should work).

https://stackoverflow.com/questions/46114214/lambda-implicit-capture-fails-with-variable-declared-from-structured-binding

* Fixes requested by Mathias

- Do not change default options in CMakeLists.txt
- Use std::optional as return value of fast_small_cut_enumeration
- Use const reference when possible