Use parallel_hashmap instead of sparsepp.
The author of sparsepp already said that this library is deprecated and
the use of 'parallel_hashmap' is preferred.

The only 'weirdness' with this is in the serializer/deserializer.
I couldn't find a way to make it work with 'streams', only with file
streams using the binary archive methods from 'parallel_hashmap'.