Allow mockturtle to run on big-endian platforms.
Most of the code is not influenced by the endianness. However the
combination of bit-filed struct and union is.

For example. Take this code:

union {
    struct
    {
        uint64_t complement : 1;
        uint64_t index : 63;
    };
    uint64_t data;
};

If we try to change the value of 'complement' using (data ^ 1), then
things will break. This commit adds compile-time macros to identify the
endianness of the platform, and change the order of 'complement' and
'index' accordingly. (Everything else is taken care by the compiler.)