Network method to create a duplicate of it.
Most (if not all) networks implemented in `mockturtle` use shared
pointers to store the internal data.  When dealing with them, the default
behavior is to share this storage them during assignments or function
calls. However, sometimes we need to make a copy of the network as well.
This commit adds a method to do exactly this.