Fix a bug when creating `pi` and using `pi_index`
Both `mig` and `xmg` uses the same logic as `aig` to retrieve a primary
input index from its node.  However, they create primary inputs
differently: instead of making their children equal to the `pi` index,
as in a `aig`, the make them equal to the numerical limit of `uint64_t`.
Effectively this means that the `pi_index` method on `mig` and `xmg` in
broken.

This commit solves the issue. Someone must check the other networks to
see if the problem also exists, e.g., `aqfp` seems to suffer from it
too.