Network method to create a duplicate. (#522)
* Network method to create a duplicate of it.

Most (if not all) networks implemented in `mockturtle` use shared
pointers to store the internal data.  When dealing with them, the default
behavior is to share this storage them during assignments or function
calls. However, sometimes we need to make a copy of the network as well.
This commit adds a method to do exactly this.

* Small fix. There is no `beginregion`.

* Tests for `clone`

* Update docs

* Fix typos

* Rename `is_clonable` to `has_clone`

* has_clone_v

* has_clone_v

* Update mig.cpp

* has_clone_v xag

* has_clone_v xmg

Co-authored-by: Siang-Yun (Sonia) Lee <siang-yun.lee@epfl.ch>