Restore tests with ASAN #647  (#673)
* Revert "Revert "Run tests with ASAN (#647)" (#672)"

This reverts commit 92ff9b2640f82762328be39632326aef1857a616.

* ci: update linux.yml to install compilers before building

Previously, the workflow assumed compilers were available in the environment,
which could cause failures. Now, compilers are explicitly installed before running CMake and Make.

* ci: update coverage.yml to use a compatible Ubuntu version

The workflow was using ubuntu-latest, which currently points to Ubuntu 24.04.
However, codecov is not yet fully compatible with Ubuntu 24.04.
This commit explicitly sets the runner to ubuntu-22.04 to ensure the coverage upload works correctly.